Allow all users to access product pages.
Updated the product controller and its associated show view so that any
user can access a product page. A condition has been added to the view,
so that only signed in users can see the product administration
controls. If a user is not signed in the they just see the product's
title, description, and image.