Remove call to launchy helper method.
Accidently left a call to the launchy helper method in one of the step
definitions. I probably should have noticed this earlier, but only saw
it when I ran the automated check-in tests. Must be more vigilant in the
future or go back to making small work in progress commits.