Hide "Sign in" link, when signed in.
https://trello.com/c/C3ZJAPJx

Previously, when an administrator was signed into the application, the
"Sign in" link was still being displayed in the application's footer,
which was causing administrator's to question wether they were signed in
or not. The footer was changed so that the "Sign in" link is only
displayed when an administrator is not signed in.

* Added a helper method for the "Sign in" link.