Show the “Add to Basket” button for each product
https://trello.com/c/h8uZLq19

Previously, you users were only able to view individual products without
any other actions, which is no good when you are trying to make some
money. An “Add to Basket” button has been added to the product list on
the “Shop” page and the homepage that allows users to add each product
to their basket.