Add a favicon
https://trello.com/c/XzNsIedE

Previously, when a user had multiple tabs open in their browser it would
be difficult for them to identify the application in their browser,
which does not help when trying to make users more aware of the
application. A favicon has been added to the public root to make the
application more visible in the users browser.