Configure the application to deploy to Heroku
https://trello.com/c/0MQaFJpB

Previously, several warnings were being thrown by Heroku when deploying
the application and assets were not being compiled correctly, which was
slowing down the deployment of the application and causing the
application to be rendered incorrectly. The Rails 12 Factor gem has been
installed and images used in stylesheets have been updated to be
included using the SASS helpers included in Rails.