Show order details
Previously, there was no page for an administrator to view the details
of an order, which was causing difficulties when an administrator needed
to see what a customer had ordered and where the order needed to be
shipped to. A page has been added that will display all of the
information about a specific order.

https://trello.com/c/ss5habYR