Add "increase quantity" button
Previously, there was no way for a customer to increase the quantity of items in
their basket, which meant they had to explicitly add a product to their basket
repeatedly to increase its quantity. Added a new button to the basket that
allows the user to increase the quantity of an item in the basket.

https://trello.com/c/dwme2BLI