Add "decrease quantity" button
Previously, there was no way for a customer to decrease the quantity of items in
their basket, which meant they had to remove the item completely before adding
the correct quantity to the basket. Added a new button to the basket that allows
the user to decrease the quantity of an item in the basket.

https://trello.com/c/TVOY3tPB