Change SimpleCov output dir while run in CircleCI
Previously, SimpleCov's test coverage report was only set up to run
locally, which meant that we couldn't get an overall view of the state
of coverage without running it locally. Updated the test suite and the
CircleCI configuration to generate a coverage report every time we
successfully run a CI build.

* Updated database_cleaner and honeybadger.

https://trello.com/c/o4KzFPee