Add a integration test for missing events
Previously, there were no tests around the 'Event' page for scenarios
where the event did not exist, which meant that we could not make any
functionality changes and be 100% confident that we had not broken any
existing functionality. Added an integration test around the 'Event'
page functionality for scenarios when the event does not exist.

https://trello.com/c/T3fE8DoF