Fix the Stripe implementation
Before, we may have had too much trust in the test suite. We hadn't
checked the application in the browser for a while, and things weren't
as rosy as we had assumed. We had not covered the old JavaScript Stripe
implementation in a feature spec. As you may have guessed, this approach
no longer worked. We fixed the order creation flow to only use the
server side.

- Added a spec helper to stub out HTTP requests to AWS S3.
- Added a product to the developer Rake task.